remove files no longer needed, fix pointers, and move methods out of plant.py
